move slider backcompat to separate file
